Moved get_invested_data to separate function
