+++ stocktracker_orchestrator/__init__.py
@@ -41,9 +41,12 @@ def orchestrator_function(context: df.DurableOrchestrationContext):
     # step 4 - Compute transactions
     stock_held = yield context.call_activity("compute_transactions", transactions)
 
+    # step 5 - Get invested data
+    invested = yield context.call_activity("get_invested_data", transactions)
+
     # Step 5 - Run main function
     result = yield context.call_activity(
-        "stocktracker", [transactions, stock_data, forex_data, stock_held]
+        "stocktracker", [stock_data, forex_data, stock_held, invested]
     )
     return [result]
